requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
sqlmodel
python-multipart

# Azure Storage SDKs
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-core==1.29.5

# Redis for caching
redis==5.0.1

# DSPy and ML dependencies
dspy-ai==3.0.3
pandas
pyee

# Image processing
Pillow

# Environment variables
python-dotenv

# Async utilities
nest-asyncio


